fix(Money): double범위를 넘은 수익률 계산에 대한 오류 수정 Math.Round()는 double로 캐스팅해와서 범위를 초과할 시 제대로 출력되지 않는 오류 발생. long으로 캐스팅 후 String.format을 사용해 수익률 정상출력

+++ src/main/java/lotto/Application.java
@@ -13,7 +13,7 @@ public class Application { // 패키지 분리
     public static void main(String[] args) {
         // TODO: 프로그램 구현
         Message.requestInputMoneyMessage();
-        int count = Money.countLotto(UserNumberScan.requestValue());
+        long count = Money.countLotto(UserNumberScan.requestValue());
         Message.lottoBuyMessage(count);
         List<Lotto> countLotto = RandomNumberCreater.generateLottoNumber(count);
         for (int j = 0; j < count; j++) {
